backend/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.13.1
appnope==0.1.3
asgiref==3.3.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==6.0.0
blinker==1.7.0
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
cloudpickle==3.0.0
comm==0.1.2
contourpy==1.0.7
cssutils==2.2.0
cycler==0.11.0
databricks-cli==0.18.0
datasets==2.11.0
debugpy==1.6.6
decorator==5.1.1
dill==0.3.6
distlib==0.3.6
Django==3.1.7
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-etc==1.3.0
django-hitcount==1.3.3
django-js-asset==1.2.2
docker==6.1.3
entrypoints==0.4
evaluate==0.4.0
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
flatbuffers==23.5.26
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.10.0
future==0.18.2
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
gunicorn==21.2.0
h5py==3.9.0
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.21.3
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.0.3
jupyter_core==5.2.0
kaggle==1.5.16
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
Mako==1.3.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mlflow==2.9.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
networkx==3.2.1
numpy==1.23.5
oauthlib==3.2.2
opendatasets==0.1.22
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
platformdirs==3.1.1
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.21.9
psutil==5.9.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.14.0
PyJWT==2.8.0
pynliner==0.8.0
pyparsing==3.0.9
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
pyscopg2==66.0.2
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2021.1
PyYAML==6.0
pyzmq==25.0.1
querystring-parser==1.2.4
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
smmap==5.0.1
soupsieve==2.2.1
SQLAlchemy==2.0.25
sqlparse==0.4.1
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
TBB==0.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.2.0
tomli==2.0.1
torch==2.1.0
torchvision==0.16.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.8.0
urllib3==1.26.15
virtualenv==20.21.0
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.15.0
xxhash==3.2.0
yarl==1.8.2
zipp==3.17.0
